refac: stat 컴포넌트 파일 이름 PascalCase 수정 및 hook dependency 수정
